Moved test files to test folder in root
