+++ client/components/Gallery.jsx
