+++ client/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App/App.jsx';
+import App from './components/App.jsx';
 
 ReactDOM.render(React.createElement(App), document.getElementById('root'));
