+++ tests/server.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-const app = require('./app.js');
+const app = require('../server/app.js');
 
 describe('test the get images path', () => {
   it('should have a status code of 200 and a defined body on response', (done) => request(app).get('/api/restaurants/1/images')
